LO CAMBIA TODO

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -47,7 +47,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setFloat(idx, comercial.getComision());
+			ps.setBigDecimal(idx, comercial.getComision());
 			return ps;
 		},keyHolder);
 		
@@ -68,7 +68,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 						 	rs.getString("nombre"),
                 						 	rs.getString("apellido1"),
                 						 	rs.getString("apellido2"),
-                						 	rs.getFloat("comisión")
+                						 	rs.getBigDecimal("comisión")
                 						 	)
         );
 		
@@ -90,7 +90,7 @@ public class ComercialDAOImpl implements ComercialDAO {
             						 						rs.getString("nombre"),
             						 						rs.getString("apellido1"),
             						 						rs.getString("apellido2"),
-            						 						rs.getFloat("comisión")) 
+            						 						rs.getBigDecimal("comisión")) 
 								, id
 								);
